remove type.h and cont.h
