+++ extlib/benz/eval.c
@@ -150,7 +150,7 @@ define_var(pic_state *pic, analyze_scope *scope, pic_value sym)
       pic_warnf(pic, "redefining variable: ~s", sym);
       return;
     }
-    pic_weak_set(pic, pic->globals, sym, pic_invalid_value());
+    pic_weak_set(pic, pic->globals, sym, pic_invalid_value(pic));
   }
 }
 
@@ -181,7 +181,7 @@ analyze_var(pic_state *pic, analyze_scope *scope, pic_value sym)
 static pic_value
 analyze_defer(pic_state *pic, analyze_scope *scope, pic_value form)
 {
-  pic_value skel = pic_cons(pic, pic_invalid_value(), pic_invalid_value());
+  pic_value skel = pic_cons(pic, pic_invalid_value(pic), pic_invalid_value(pic));
 
   pic_set_car(pic, scope->defer, pic_cons(pic, pic_cons(pic, form, skel), pic_car(pic, scope->defer)));
